app: basic dependency navigation
When highlighting data dependencies upstream instructions may be "far
away" withing the instruction stream. In such cases the user would
have to scroll for a while and not miss the dependency while doing
so. Also, there was no visual indicator that there even is a
dependency, if it's not immediately visible.

Implement a "dependency-arrows" column in the log view, imeediately to
the left of the instructions stream. It's ASCII-based lines between
collected upstream dependencies of the selected line. For example:

  ┌─ r1 = *(u64 *)(r0 -8)
  │  r2 = 0xff6f3b1a00e97010
  └─ *(u64 *)(r2 +8) = r1

The vertical lines are interactive: if one of the closest anchored
lines is not visible, it changes to an arrow on hover, and scrolls to
the closes relevant line on click.

The lines are updated on each updateView() and are visible only if
there is a selected mem slot with known dependencies. Navigation
targets are computed on hover at "dependency-arrows" column.

The memSlotDependencies field of the state is now stored as a sorted
array of indices, and not as a Set. This is useful when computing
navigation targets.

Closes: https://github.com/libbpf/bpfvv/issues/10
Signed-off-by: Ihor Solodrai <isolodrai@meta.com>