app: don't show independent writes
When collecting dependencies for a mem slot, so far the app always
showed the chain of "last known writes". So, for example:

  11: r1 = 1
  ...
  17: r1 = 2

By clicking on r1 at line 17 the line 11 would be highlighted. This is
not a real dependency and not particularly useful.

Rewrite collectMemSlotDependencies routine to more intuitive
implementation using a recursive helper `memSlotDependencies`. The
helper takes line idx and slot id and then looks at the instruction at
last known write and its effect to determine whether it's appropriate
to traverse further.

The design hasn't changed: dependency highlighting stops at the first
instruction with multiple slots read, with the exception of a selected
mem slot update.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>